+++ odoo/local-src/specific_stock/wizards/__init__.py
@@ -2,3 +2,4 @@ from . import stock_receive_lot
 from . import put_in_pack_helper
 from . import generate_voice_identifier
 from . import open_uop
+from . import stock_picking_return